+++ archived_pipelines/basic_rag/pipeline_vector.py
@@ -49,10 +49,10 @@ class BasicRAGPipelineVector:
                     doc_id,
                     title,
                     text_content,
-                    VECTOR_COSINE(document_embedding_vector, TO_VECTOR('{query_embedding_str}', 'DOUBLE')) as similarity_score
+                    VECTOR_COSINE(document_embedding_vector, TO_VECTOR('{query_embedding_str}', 'FLOAT')) as similarity_score
                 FROM RAG.SourceDocuments
                 WHERE document_embedding_vector IS NOT NULL
-                AND VECTOR_COSINE(document_embedding_vector, TO_VECTOR('{query_embedding_str}', 'DOUBLE')) > {similarity_threshold}
+                AND VECTOR_COSINE(document_embedding_vector, TO_VECTOR('{query_embedding_str}', 'FLOAT')) > {similarity_threshold}
                 ORDER BY similarity_score DESC
             """
